line chart: refactor to separate internal types
There are few types that are used internal to the charting library and
there are few public ones. Since the public one is a subset of the
internal one, we are turning `types`, which is too generic, into
`internal_types` and selectively re-export types in the `public_types`.
