+++ tensorboard/webapp/widgets/line_chart_v2/lib/internal_types.ts
