+++ tensorboard/webapp/widgets/line_chart_v2/lib/utils.ts
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-import {Extent, Polyline, Rect} from './types';
+import {Extent, Polyline, Rect} from './internal_types';
 
 export function convertRectToExtent(rect: Rect): Extent {
   return {
